STY: Put pure data mappings in separate files (#1005)
Created a private codecs module
